Split stage pipeline tests by responsibility
Move the remaining stage artefact pipeline tests into focused modules for
core staging, cargo-binstall archives, PowerShell sidecars, and path-safety
coverage. Keep the shared strategies and helper builders in `conftest.py`
so each test module can stay narrow.

Delete the now-empty `test_pipeline.py` module after migrating its test
coverage.
